Rejig RegionProviderList initialisation process, move to Maps/ as per @kring's excellent suggestion.
