+++ lib/Map/RegionProvider.js
@@ -8,7 +8,7 @@ var loadText = require('terriajs-cesium/Source/Core/loadText');
 var when = require('terriajs-cesium/Source/ThirdParty/when');
 
 var corsProxy = require('../Core/corsProxy');
-var ModelError = require('./ModelError');
+var ModelError = require('../Models/ModelError');
 var xml2json = require('../ThirdParty/xml2json');
 var URI = require('URIjs');
