+++ lib/Map/RegionProviderList.js
@@ -28,77 +28,48 @@ var RegionProvider = require('./RegionProvider');
   provider for a given dataset.
 */
 
-var RegionProviderList = function(terria) {
-    /** 
-     * URL pointing to a JSON file where the region provider definitions will be loaded from. 
-     * @type {String}
-     */
-    this.terria = terria;
-
+var RegionProviderList = function() {
     /** 
      * List of RegionProviders, once loaded from file. 
      * @type {RegionProvider[]} 
      */
     this.regionProviders = [];
 
-    /**
-     * Has this list been initialized from file yet.
-     * @type {Boolean}
-     */
-    this._initialized = false;
-
 };
+/* A static dictionary of promises to RegionProviderLists */
+RegionProviderList.metaList = {};
 
 /**
- * Fetch JSON file and create ordered list of region providers from it. 
- * @returns Promise.
+ * Returns a promise for a RegionProviderList instantiated from this url. Previous loads
+ * are cached.
  */
-RegionProviderList.prototype.init = function(url) {
-    if (this._initialized) {
-        return when(this);
+RegionProviderList.fromUrl = function(url) {
+   if (!defined(RegionProviderList.metaList[url])) {
+        RegionProviderList.metaList[url] = loadJson(url)
+            .then(function(o) { 
+                return new RegionProviderList().initFromObject(o);
+            });
     }
-    if (!defined(url)) {
-        url = this.terria.regionMappingDefinitionsUrl;
-    }
-
-    var that = this;
-    return loadJson(url)
-        .then(function(o) { 
-            return that.initFromObject(o);
-        });
+    return RegionProviderList.metaList[url];
 };
 
 /**
  * Initialises from the already-retrieved contents of a JSON file.
  */
 RegionProviderList.prototype.initFromObject = function(obj) {
-    if (this._initialized) {
-        throw new DeveloperError ("RegionProviderList already initialized.");
-    }
     var that = this;
     Object.keys(obj.regionWmsMap).forEach(function(r) {
         that.regionProviders.push(new RegionProvider(r, obj.regionWmsMap[r], that));
     });
-    this._initialized = true;
     return this;
 };
 
-/**
- * Has RegionProviderList finished loading config file?
- * @returns {Boolean}
- */
-RegionProviderList.prototype.ready = function() {
-    return this._initialized;
-};
 
 /**
   *  Find what kind of region-mapped dataset this is by exhaustively looking for every alias in every column. 
   *  @param varNames {String[]} Array of names of variables to choose amongst.
   */
 RegionProviderList.prototype.chooseRegionProvider = function (varNames) {
-    if (!this._initialized) {
-        throw new DeveloperError('Region provider list not initialized.');
-    }
     //try to figure out the region variable and type based on aliases
     var i;
     for (i = 0; i < this.regionProviders.length; i++) {
@@ -121,9 +92,6 @@ RegionProviderList.prototype.chooseRegionProvider = function (varNames) {
     @return {Object} Region Provider, or null.
 */
 RegionProviderList.prototype.getRegionProvider = function(regionType) {
-    if (!this._initialized) {
-        throw new DeveloperError('Region mapping list not initialized.');
-    }
     var r = this.regionProviders.filter(function(p) { return p.regionType.toLowerCase() === regionType.toLowerCase(); });
     if (r.length > 1)
         throw new DeveloperError ('More than one definition of region provider: ' + regionType);
