+++ lib/Models/CsvCatalogItem.js
@@ -32,7 +32,7 @@ var TileLayerFilter = require('../ThirdParty/TileLayer.Filter');
 var CsvDataset = require('./CsvDataset');
 var CsvVariable = require('./CsvVariable');
 
-var RegionProviderList = require('./RegionProviderList');
+var RegionProviderList = require('../Map//RegionProviderList');
 var ImageryProviderHooks = require('../Map/ImageryProviderHooks');
 
 
@@ -586,10 +586,8 @@ function loadTable(csvItem, text) {
         return;
     }
     console.log('No lat&lon columns found in csv file - trying to match based on region');
-    // TODO Consider preserving the regionmapping object for future loads?
-    var rm = new RegionProviderList(csvItem.terria);
-    return rm.init()
-        .then(function() {
+    return RegionProviderList.fromUrl(csvItem.terria.regionMappingDefinitionsUrl)
+        .then(function(rm) {
             if (dataset.checkForRegionVariable(rm)) {
                 return updateTableStyle(csvItem, initRegionMapStyle(csvItem, rm));
             }
@@ -597,7 +595,7 @@ function loadTable(csvItem, text) {
         .then(function() {
             if (csvItem._regionMapped) {
                 finishTableLoad(csvItem);
-                return rm;
+                return;
             }
             throw new ModelError({
                 sender: csvItem,
